perf(overlay): resolve each incumbent owner once per flip, not once per row
The import asked mirror_user_map who owned every single record it
landed. An estate's owners number in the dozens where its records
number in the hundreds of thousands, so the run spent one query per
record re-answering a question it had already answered. Live lookups
are now memoized for the run, which is safe because the mirror is
frozen while it executes.

Owner resolution moves to its own file as it goes: it decides whether a
flipped record is visible to one seat or to the whole workspace, which
is a security decision rather than a mapping one, and flipwriters.go
was over the length cap regardless.

The malformed-bundle refusal still echoed object_class, which is
untrusted bundle content like the rest of the row. It now names only
which identifier is absent, which was always the whole diagnosis.

Corrects a lifecycle assertion's message that still said "want 2" after
the expectation moved to 3.
